* master: (26 commits)
  Fixed test name.
  Bumped podspec and SDK version.
  Fixed some threading & iOS 7 bugs
  Fix for crash on iOS 7.
  Only hide title instead of the whole title bar.
  Fix for crash because of deleted templates.
  If OLKiteViewController is initialised with an OLPrintOrder (launch to shipping) then hide the title bar so as to avoid fast animations looking weird.
  Dispatching transitioning to the next screen to the main thread.
  Removed background threading suspected to cause problems.
  Added the analytics call to the completion handler as well.
  Set up AB test variants after we are done fetching the plist.
  Fixed not being able to dismiss shipping.
  Specified Skylab version.
  Analytics
  AB Test showing and not showingProductDescription.
  Launching with a printOrder now shows product description screen before shipping.
  Done button on receipt screen.
  Bumped SDK version
  Bumped SDK version
  Only fetch api key-specific plist, only fall back to default if it does not exist.
  ...

+++ Kite SDK/PSPrintSDK/OLPhotoSelectionViewController.m
@@ -663,7 +663,7 @@ static const NSUInteger kTagAlertViewSelectMorePhotos = 99;
         NSInteger qty = [self.userSelectedPhotos[imageIndex] extraCopies];
         if (qty > 0 && self.product.productTemplate.templateUI != kOLTemplateUIFrame && self.product.productTemplate.templateUI != kOLTemplateUIPhotobook && self.product.productTemplate.templateUI != kOLTemplateUIPoster){
             qtyLabel.hidden = NO;
-            qtyLabel.text = [NSString stringWithFormat:@"%ld", qty+1];
+            qtyLabel.text = [NSString stringWithFormat:@"%ld", (long)qty+1];
         }
         else{
             qtyLabel.hidden = YES;
